Fixes indentation for use with `david` tool from `david-dm`.
`grunt-cli-mocha` is using
[david-dm](https://david-dm.org/Rowno/grunt-mocha-cli) to dependency
updates. There exists an npm package
[david](https://npmjs.org/package/david) that checks project
dependencies for updates via the command `david update` and writes to
`package.json`.

Previous to this commit, use of this tool would generate a meaningless
diff due to the nonstandard `package.json` indentation in the
`grunt-cli-mocha` project.

This commit regenerates `package.json` with david-compatible indentation
for friendlier diffs.